emspro/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
mysqlclient==2.1.1
gunicorn
django-filter